requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.40.0

# Tokenizers and model utilities
sentencepiece>=0.2.0
tokenizers>=0.19.0

# HuggingFace Hub
huggingface-hub>=0.23.0

# Docker and evaluation
docker>=7.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# Logging and monitoring
wandb>=0.16.0
tensorboard>=2.16.0

# API providers (for inference without GPUs)
openai>=1.0.0
anthropic>=0.25.0
together>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# For Modal (optional)
modal>=0.62.0

# Git operations
GitPython>=3.1.0
